Move new criteria eval functions to new file.
