+++ lib/math-trig.js
@@ -1,9 +1,10 @@
-var utils = require('./utils');
+var _ = require('./limited-lodash');
+var criteriaEval = require('./criteria-eval');
 var error = require('./error');
-var statistical = require('./statistical');
 var information = require('./information');
 var matrix = require('./matrix');
-var _ = require('./limited-lodash');
+var statistical = require('./statistical');
+var utils = require('./utils');
 
 exports.ABS = function(number) {
   number = utils.parseNumber(number);
@@ -1058,7 +1059,7 @@ exports.SUM = function () {
  */
 exports.SUMIF = function () {
   var sum = 0;
-  utils.performIf(arguments, function(value) {
+  criteriaEval.performIf(arguments, function(value) {
     if (typeof value === 'number')
       sum += value;
   });
@@ -1067,7 +1068,7 @@ exports.SUMIF = function () {
 
 exports.SUMIFS = function () {
   var sum = 0;
-  utils.performIfs(arguments, function (value) {
+  criteriaEval.performIfs(arguments, function (value) {
     if (typeof value === 'number')
       sum += value;
   });
