+++ lib/statistical.js
@@ -1,9 +1,10 @@
+var _ = require('./limited-lodash');
+var criteriaEval = require('./criteria-eval');
+var error = require('./error');
+var jStat = require('jStat').jStat;
 var mathTrig = require('./math-trig');
 var text = require('./text');
-var jStat = require('jStat').jStat;
 var utils = require('./utils');
-var error = require('./error');
-var _ = require('./limited-lodash');
 
 var SQRT2PI = 2.5066282746310002;
 
@@ -66,7 +67,7 @@ exports.AVERAGEA = function () {
 exports.AVERAGEIF = function () {
   var sum = 0;
   var count = 0;
-  utils.performIf(arguments, function(value) {
+  criteriaEval.performIf(arguments, function(value) {
     if (typeof value === 'number') {
       sum += value;
       count++;
@@ -80,7 +81,7 @@ exports.AVERAGEIF = function () {
 exports.AVERAGEIFS = function () {
   var sum = 0;
   var count = 0;
-  utils.performIfs(arguments, function (value) {
+  criteriaEval.performIfs(arguments, function (value) {
     if (typeof value === 'number') {
       sum += value;
       count++;
@@ -435,14 +436,14 @@ exports.COUNTBLANK = function (range) {
  */
 exports.COUNTIF = function () {
   var count = 0;
-  utils.performIf(arguments, function() { ++count; });
+  criteriaEval.performIf(arguments, function() { ++count; });
   return count;
 };
 
 exports.COUNTIFS = function () {
   var count = 0;
   var args = [arguments[0]].concat(utils.argsToArray(arguments));
-  utils.performIfs(args, function () { ++count; });
+  criteriaEval.performIfs(args, function () { ++count; });
   return count;
 };
